# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem booking lapangan futsal.",
        "title": "GoFutsal API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Developer API Support",
            "url": "https://github.com/HenryKristofani",
            "email": "henry@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/courts": {
            "get": {
                "description": "Menampilkan semua lapangan futsal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courts"
                ],
                "summary": "Get all courts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Court"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan data lapangan futsal baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courts"
                ],
                "summary": "Tambah lapangan baru",
                "parameters": [
                    {
                        "description": "Court Data",
                        "name": "court",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Court"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Court"
                        }
                    }
                }
            }
        },
        "/api/courts/{id}": {
            "get": {
                "description": "Menampilkan detail lapangan futsal berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courts"
                ],
                "summary": "Get court by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Court ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Court"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui data lapangan futsal berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courts"
                ],
                "summary": "Update court",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Court ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Court Data",
                        "name": "court",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Court"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data lapangan futsal berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courts"
                ],
                "summary": "Delete court",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Court ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Court": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "integer"
                }
            }
        },
...
    }
}